Day 3 End
Fixed Fire Spreading, before it wasn't working correctly.
Changed Forest from a 1d array to a 2d array, also made it easy to change number of trees.
Fixed a bug that wouldn't permit savedTrees to be re-burnt
Made it so that when you save a tree, the eight trees around it are also saved